// src/app/weather.interface.ts
export interface Weather {
    lat?: any;
    lon?: any;
    name?: string;
    country?: string;
    date?: any;
    temp?: any;
    temp_max?: number;
    temp_min?: number;
    humidity?: number;
    description?: string;
    feels_like?: number;
    icon?: string;
    sunrise?: any;
    sunset?: any;
    imgTitle?: any;
    forecast?: any;
}